requirements.txt
aiogram>=3.0,<4.0
openai>=1.0
chromadb>=0.4.0
pypdf>=3.0
python-docx>=1.0
python-dotenv>=1.0
